+++ src/pages/login/index.ts
@@ -1,1 +1,1 @@
-export * from './login';
+export { Login } from './login-container';
